--- src/Components/Produtos/CardProduto.js
@@ -1,32 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import produtos from "../produtos.json"
-
-const ProdContainer = styled.div`
-  border: 1px solid gray;
-  width: 200px;
-  margin: 10px;
-  padding: 10px;
-  align-items: center;
-`;
-
-const ImgCard = styled.img`
-  width: 100%;
-  text-align: center;
-`;
-
-
-export function CardProduto(props) {
-
-
-  return (
-    <ProdContainer key={props.produto.id}>
-      <ImgCard src={props.produto.imagem} />
-      <b name={props.produto.nome}>Produto: {props.produto.nome} </b>
-      <p valor={props.produto.valor}>R$: {props.produto.valor},00</p>
-      <button>Adicionar ao carrinho</button>
-    </ProdContainer>
-  )
-
-}
-
